requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pathlib
pymysql==1.1.0
python-dotenv==1.0.0
sqlalchemy-stubs
